+++ src/simpleprop.cpp
@@ -15,33 +15,16 @@ SimpleProp::SimpleProp()
 // Copy constructor
 SimpleProp::SimpleProp( const SimpleProp& rhs )
 {
-	SimpleProp::copy( rhs ); // use the copy utility
+	OneHiddenNet::copy( rhs ); // use the copy utility
 }
 
 // Overloaded = operator
 SimpleProp& SimpleProp::operator= ( const SimpleProp& rhs )
 {
 	if ( &rhs != this ) // check for self-assignment
-		SimpleProp::copy( rhs ); // use the copy utility
-	
-	return *this; // enables A = B = C
-}
+		OneHiddenNet::copy( rhs ); // use the copy utility
 
-// Copy utility
-void SimpleProp::copy( const SimpleProp& rhs )
-{
-	Network::copy( rhs ); // call immediate base object copy
-	nHidden = rhs.nHidden;
-	hW = rhs.hW;
-	hWup = rhs.hWup;
-	y = rhs.y;
-	I = rhs.I;
-	hO = rhs.hO;
-	oW = rhs.oW;
-	h_err = rhs.h_err;
-	o_err = rhs.o_err;
-	hG = rhs.hG;
-	oG = rhs.oG;
+	return *this; // enables A = B = C
 }
 
 // Loads a DataSet object into the SimpleProp Model
@@ -259,153 +242,6 @@ vector< double > SimpleProp::hiddenSaliency()
 unsigned SimpleProp::df()
 {
 	return ( ( ( theData.getInput() + 2 ) * nHidden ) + 1 );
-}
-
-// Randomizes initial weights in this Network object
-void SimpleProp::randomize()
-{
-	assert( builtFlag ); // the object must have been built first
-	
-	// Randomize the weights
-	hW.random( randomLimit ); // randomize hW
-	nvec::random( oW, randomLimit ); // randomize oW
-
-	// Log to history file
-	if ( historyFlag ) // make sure flag for history is set
-	{
-		// Construct the message to the output stream
-		ostringstream fileStream;
-		fileStream << endl << "I've randomized the weights of a " << objType << " object."
-			<< endl << endl;
-		addHistory( fileStream ); // append to history file
-	}
-
-	weightsSetFlag = true; // set flag to indicate weights now set
-}
-
-// Save this Network object to a file, takes filename as ( string ) argument
-//    returns boolean flag to indicate if file succesfully saved
-bool SimpleProp::save( string& filename )
-{
-	bool success = false; // flag to indicate file successfully saved
-	
-	// Open the output file to save data, overwrite file if it exists
-	ofstream savefile( filename.c_str(), ios::out | ios::trunc );
-	
-	// Test to insure it was opened
-	if ( !savefile.is_open() )
-		util::screen() << "Error in opening file to save " << objType << " network!" << endl;
-	else
-	{
-		outputHeader( savefile ); // output the header to the file
-
-		// Next line is hidden layer label, followed by hidden layer matrix
-		savefile << "Hidden layer:" << endl;
-		savefile << hW.setHeader( false ); // no header is necessary
-
-		// Followed by output layer label, followed by output layer vector
-		savefile << "Output layer:" << endl;
-		savefile << oW << endl;
-
-		// Print message to user notifying successful save to file
-		util::screen() << "The " << objType << " network was successfully saved to " << filename;
-		util::screen() << "." << endl;
-
-		savefile.close(); // close output file
-
-		success = true; // set flag to indicate file successfully saved
-
-		// Log to history file
-		if ( historyFlag ) // make sure flag for history is set
-		{
-			// Construct the output stream with object name and header
-			ostringstream fileStream;
-			fileStream << "I've saved the file " << filename << ":" << endl;
-			outputHeader( fileStream ); // output the header to the history file
-			fileStream << endl;
-			
-			addHistory( fileStream ); // append to history file
-		}	
-	}
-
-	return success; // return flag to indicate if file successfully saved
-}
-
-// Load this Network object from a file, takes filename as ( string ) argument
-//    returns boolean flag to indicate if file succesfully loaded
-bool SimpleProp::load( string& filename )
-{
-	// Easier on the eyes
-	unsigned nInput = theData.getInput(); // number of input nodes
-	
-	bool success = false; // flag to indicate file successfully loaded
-	
-	string lineString, // holder for strings pulled from file
-		goodFile; // the name of the actual file
-
-	// Open the file as read-only, and associate it with the 'label' loadfile
-	goodFile = util::getGoodFile( filename );
-	ifstream loadfile( goodFile.c_str(), ios::in );
-
-	getline( loadfile, lineString ); // 1st line should be SimpleProp
-	util::chopEndl( lineString ); // remove <cr> from end of string if exists
-	assert( lineString == objType );
-	
-	getline( loadfile, lineString ); // "Bias nodes on all layers by definition"
-
-	unsigned nInputFromFile; // number of input nodes obtained from file
-	loadfile >> nInputFromFile; // retrieve number of input nodes
-	getline( loadfile, lineString ); // " input nodes"
-
-	// Make sure number of input nodes matches dataset
-	if ( nInputFromFile != nInput )
-	{
-		util::screen() << "I cannot load this file:" << endl;
-		util::screen() << "The number of input nodes do not match the dataset ( ";
-		util::screen() << nInput << " )" << endl;
-	}
-	else
-	{
-		getline( loadfile, lineString ); // "1 hidden layer by definition"
-
-		loadfile >> nHidden; // retrieve number of hidden nodes
-		getline( loadfile, lineString ); // " hidden nodes"
-		setHidden( nHidden ); // set the architecture of this object
-
-		getline( loadfile, lineString ); // "1 output node by definition"
-
-		getline( loadfile, lineString ); // "Hidden layer:"
-		loadfile >> hW; // retrieve hidden layer
-
-		// Pick up the space and carriage return after the Matrix object
-		getline( loadfile, lineString ); // THIS IS CRITICAL!
-
-		getline( loadfile, lineString ); // "Output layer:"
-		loadfile >> oW; // retrieve output layer
-
-		weightsSetFlag = true; // set flag to indicate weights now set
-
-		outputHeader( util::screen() ); // report to user
-		util::screen() << "I've loaded the file." << endl;
-
-		// Log to history file
-		if ( historyFlag ) // make sure flag for history is set
-		{
-			// Construct the output stream with object name and header
-			ostringstream fileStream;
-			fileStream << "I've loaded the file " << goodFile << ":" << endl;
-			outputHeader( fileStream ); // output header to history file
-			fileStream << endl;
-			
-			addHistory( fileStream ); // append to history file
-		}
-
-		success = true; // set flag to indicate file successfully loaded
-	}
-
-	loadfile.close(); // close input file
-
-	return success; // return flag to indicate if file successfully loaded
 }
 
 // Outputs a header to ostream describing this SimpleProp model architecture
@@ -739,16 +575,6 @@ void SimpleProp::removeInputs( const vector< unsigned >& v )
 	hG = hG.excludecols( v ); // remove hidden gradient corresponding to inputs
 
 	hWup.resize( nHidden, theData.getInput() + 1 ); // resize hidden weight update Matrix
-}
-
-// Convert weight gradient structure to single vector stackG
-void SimpleProp::pack()
-{
-	// Start by converting hidden gradients Matrix to stackG
-	stackG = hG.toVector();
-
-	// Then append output gradients vector to stackG
-	std::copy( oG.begin(), oG.end(), back_inserter( stackG ) );
 }
 
 // Convert single vector stackG back to weight gradient structure